Implement tonic metrics proto transformations (#1184)
