+++ opentelemetry-proto/src/transform/common.rs
@@ -1,7 +1,7 @@
-#[cfg(any(feature = "trace", feature = "logs"))]
+#[cfg(any(feature = "trace", feature = "logs", feature = "metrics"))]
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
-#[cfg(any(feature = "trace", feature = "logs"))]
+#[cfg(any(feature = "trace", feature = "logs", feature = "metrics"))]
 pub(crate) fn to_nanos(time: SystemTime) -> u64 {
     time.duration_since(UNIX_EPOCH)
         .unwrap_or_else(|_| Duration::from_secs(0))
