+++ opentelemetry-proto/src/transform/logs.rs
@@ -121,7 +121,11 @@ pub mod tonic {
                     attributes: resource_attributes(&log_data.resource).0,
                     dropped_attributes_count: 0,
                 }),
-                schema_url: "".to_string(),
+                schema_url: log_data
+                    .resource
+                    .schema_url()
+                    .map(Into::into)
+                    .unwrap_or_default(),
                 scope_logs: vec![ScopeLogs {
                     schema_url: log_data
                         .instrumentation
@@ -268,7 +272,11 @@ pub mod grpcio {
                     dropped_attributes_count: 0,
                     ..Default::default()
                 }),
-                schema_url: "".to_string(),
+                schema_url: log_data
+                    .resource
+                    .schema_url()
+                    .map(Into::into)
+                    .unwrap_or_default(),
                 scope_logs: RepeatedField::from_vec(vec![ScopeLogs {
                     schema_url: log_data
                         .instrumentation
